requirements.txt
streamlit
pandas
SQLAlchemy
alembic
python-dotenv
psycopg2-binary
